+++ tools/tcl/rbbram/util.tcl
@@ -1,6 +1,6 @@
-# $Id: util.tcl 603 2014-11-09 22:50:26Z mueller $
+# $Id: util.tcl 883 2017-04-22 11:57:38Z mueller $
 #
-# Copyright 2011-2014 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
+# Copyright 2011-2017 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 #
 # This program is free software; you may redistribute and/or modify it under
 # the terms of the GNU General Public License as published by the Free
@@ -13,6 +13,7 @@
 #
 #  Revision History:
 # Date         Rev Version  Comment
+# 2017-04-22   883   2.0.1  setup: now idempotent
 # 2014-11-09   603   2.0    use rlink v4 address layout
 # 2011-03-19   372   1.0    Initial version
 #
@@ -28,6 +29,7 @@ namespace eval rbbram {
   # setup: amap definitions for rbd_bram
   # 
   proc setup {base} {
+    if {[rlc amap -testname br.cntl $base]} {return ""}
     rlc amap -insert br.cntl [expr {$base + 0x00}]
     rlc amap -insert br.data [expr {$base + 0x01}]
   }